fix open deltachat from uri doesn't detect the uri on macOS closes https://github.com/deltachat/deltachat-desktop/issues/2257 show qr code content if decoding failed also add a workaround for https://github.com/deltachat/deltachat-core-rust/issues/1969 until core can accept both.
