+++ .gitignore
@@ -82,3 +82,4 @@ src/shared/build-info.ts
 # ignore provisioning file
 embedded.provisionprofile
 
+stub.js
